+++ hyp3_sdk/asf_search.py
@@ -2,7 +2,7 @@ from typing import Iterable, List, Union
 
 import requests
 
-_BASLINE_API = 'https://api.daac.asf.alaska.edu/services/search/baseline'
+_BASELINE_API = 'https://api.daac.asf.alaska.edu/services/search/baseline'
 _SEARCH_API = 'https://api.daac.asf.alaska.edu/services/search/param'
 
 
@@ -50,7 +50,7 @@ def get_nearest_neighbors(granule: str, max_neighbors: int = 2,) -> List[dict]:
         'output': 'jsonlite',
         'master': granule,
     }
-    response = requests.get(_BASLINE_API, params=params)
+    response = requests.get(_BASELINE_API, params=params)
     response.raise_for_status()
     all_neighbors = response.json()['results']
     selected_neighbors = [n for n in all_neighbors if n['temporalBaseline'] < 0]