+++ hyp3_sdk/exceptions.py
@@ -5,9 +5,5 @@ class HyP3Error(Exception):
     """Base Exception for Hyp3_sdk"""
 
 
-class ValidationError(HyP3Error):
-    """Raise when jobs do not pass validation"""
-
-
 class AuthenticationError(HyP3Error):
     """Raise when authentication does not succeed"""